requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aniso8601==8.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bitarray==3.0.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==7.1.2
colorama==0.4.6
cytoolz==1.0.0
distlib==0.3.9
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
filelock==3.16.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
pipenv==2024.3.1
platformdirs==4.3.6
propcache==0.2.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-telegram-bot==21.7
pytz==2020.1
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
setuptools==75.3.0
six==1.15.0
sniffio==1.3.1
SQLAlchemy==1.3.18
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
web3==7.4.0
websockets==13.1
Werkzeug==1.0.1
yarl==1.17.1
